Add files via upload
The purpose of this project is to create a machine learning model capable of classifying emails as spam or not spam. This is achieved by using the Multinomial Naive Bayes algorithm, which is effective for text classification tasks. The model is trained using a dataset of non-spam emails and evaluated based on its ability to correctly predict whether a given email is spam or not.